build-agent: chore: extract git ops to adws/vcs module
