+++ adws/triggers/trigger_webhook.ts
@@ -12,10 +12,10 @@ import * as http from 'http';
 import { log, PullRequestWebhookPayload, allocateRandomPort, isPortAvailable, getTargetRepoWorkspacePath, rebuildProjectCostCsv } from '../core';
 import { fetchExchangeRates } from '../core/costReport';
 import { costCommitQueue } from '../core/costCommitQueue';
-import { commitAndPushCostFiles, pullLatestCostBranch } from '../github/gitOperations';
+import { commitAndPushCostFiles, pullLatestCostBranch } from '../vcs';
 import { isActionableComment, isClearComment, isAdwRunningForIssue, truncateText, getRepoInfoFromPayload, getRepoInfo } from '../github';
 import { clearIssueComments } from '../adwClearComments';
-import { removeWorktreesForIssue } from '../github/worktreeOperations';
+import { removeWorktreesForIssue } from '../vcs';
 import { handlePullRequestEvent, wasMergedViaPR } from './webhookHandlers';
 import { validateWebhookSignature } from './webhookSignature';
 import { checkIssueEligibility } from './issueEligibility';
